account: fold Security Keys into Access Control; drop /account page
The split between /account (Security keys only) and /users (Access
Control — users / tokens / SSO) was an artefact of how I'd scoped
PR #1 small, not a deliberate design call. Security keys are an
authentication factor: same access-control bucket as the WebUI
users, API tokens, and SSO config that already live on the
/users page. Two thin pages with overlapping scopes was the wrong
shape.

Changes
  - Port the Security Keys section into /users between WebUI Users
    and System Users. Same RPC calls, same simplewebauthn-driven
    registration flow, same origin-precheck banner (IP / wrong
    hostname / non-HTTPS).
  - Load `auth.webauthn.list` + `auth.webauthn.config` alongside
    the rest of /users's data in `refresh()`.
  - Convert the delete confirmation to use NASty's existing
    `confirm(title, message, opts)` helper instead of the inline
    `window.confirm`.
  - Delete `webui/src/routes/account/+page.svelte`.
  - Drop the `/account` entry from the layout nav + the
    `KeyRound` icon import that only existed for it. Merge the
    security-key search keywords (passkey, yubikey, touch id,
    webauthn, fido, 2fa, mfa) into /users's keyword list so the
    sidebar search still finds the section by name.

The login-shortcut flow (PR #2) and any future self-service
splits (password change, OIDC link) can grow out of /users in
place — fewer migrations needed if the page name "Access Control"
stays the natural home for all auth-factor management.
