+++ webui/src/routes/+layout.svelte
@@ -62,7 +62,6 @@
 		ScrollText,
 		Search,
 		AlertTriangle,
-		KeyRound,
 	} from '@lucide/svelte';
 	import { goto } from '$app/navigation';
 	import { refreshState } from '$lib/refresh.svelte';
@@ -568,7 +567,6 @@
 				{ href: '/logs',     label: 'Logs',           icon: ScrollText },
 				{ href: '/update',   label: 'Update',         icon: RefreshCw },
 				{ href: '/users',    label: 'Access Control', icon: ShieldCheck },
-				{ href: '/account',  label: 'Account',        icon: KeyRound },
 				{ href: '/settings', label: 'Settings',       icon: Settings },
 			]},
 		];
@@ -637,8 +635,7 @@
 		{ href: '/hardware', label: 'Hardware', keywords: ['hardware', 'pci', 'iommu', 'group', 'passthrough', 'vfio', 'gpu', 'device', 'driver', 'lspci'] },
 		{ href: '/logs', label: 'Logs', keywords: ['log', 'journal', 'systemd', 'debug', 'error', 'follow', 'stream'] },
 		{ href: '/update', label: 'Update', keywords: ['update', 'upgrade', 'version', 'release', 'nixos', 'rebuild', 'generation'] },
-		{ href: '/users', label: 'Access Control', keywords: ['user', 'password', 'role', 'admin', 'group', 'permission', 'token', 'api', 'access', 'auth', 'login'] },
-		{ href: '/account', label: 'Account', keywords: ['account', 'security key', 'webauthn', 'passkey', 'yubikey', 'touch id', 'windows hello', 'authenticator', 'fido', '2fa', 'mfa'] },
+		{ href: '/users', label: 'Access Control', keywords: ['user', 'password', 'role', 'admin', 'group', 'permission', 'token', 'api', 'access', 'auth', 'login', 'security key', 'webauthn', 'passkey', 'yubikey', 'touch id', 'windows hello', 'authenticator', 'fido', '2fa', 'mfa'] },
 		{ href: '/settings', label: 'Settings', keywords: ['setting', 'hostname', 'timezone', 'clock', 'network', 'ip', 'dhcp', 'dns', 'bond', 'vlan', 'notification', 'email', 'smtp', 'telegram', 'webhook', 'tuning', 'nfs threads', 'metrics', 'prometheus', 'telemetry', 'log level'] },
 	];
